cl8-api/requirements.txt
flask
python-dotenv
requests
flask-oauthlib
python-jose
airtable-python-wrapper

pytest
pytest-watch
